+++ src/libraries/Native/Unix/System.Globalization.Native/pal_timeZoneInfo.h
@@ -20,6 +20,6 @@ typedef enum
     TimeZoneDisplayName_ExemplarCity = 4,
 } TimeZoneDisplayNameType;
 
-PALEXPORT int32_t GlobalizationNative_WindowsIdToIanaId(const UChar* windowsId, const UChar* region, UChar* ianaId, int32_t ianaIdLength);
+PALEXPORT int32_t GlobalizationNative_WindowsIdToIanaId(const UChar* windowsId, const char* region, UChar* ianaId, int32_t ianaIdLength);
 PALEXPORT int32_t GlobalizationNative_IanaIdToWindowsId(const UChar* ianaId, UChar* windowsId, int32_t windowsIdLength);
 PALEXPORT ResultCode GlobalizationNative_GetTimeZoneDisplayName(const UChar* localeName, const UChar* timeZoneId, TimeZoneDisplayNameType type, UChar* result, int32_t resultLength);
