+++ src/libraries/System.Private.CoreLib/src/System/TimeZoneInfo.MinimalGlobalizationData.cs
@@ -28,5 +28,17 @@ namespace System
             idIsIana = false;
             return null;
         }
+
+        private static unsafe bool TryConvertIanaIdToWindowsId(string ianaId, bool allocate, out string? windowsId)
+        {
+            windowsId = null;
+            return false;
+        }
+
+        private static unsafe bool TryConvertWindowsIdToIanaId(string windowsId, string? region, bool allocate,  out string? ianaId)
+        {
+            ianaId = null;
+            return false;
+        }
     }
 }
