+++ src/libraries/System.Runtime/tests/System/TimeZoneInfoTests.cs
@@ -2682,7 +2682,9 @@ namespace System.Tests
         [ConditionalTheory(nameof(SupportIanaNamesConversion))]
         [InlineData("Pacific Standard Time", "America/Vancouver", "CA")]
         [InlineData("Pacific Standard Time", "America/Los_Angeles", "US")]
+        [InlineData("Pacific Standard Time", "America/Los_Angeles", "\u0600NotValidRegion")]
         [InlineData("Central Europe Standard Time", "Europe/Budapest", "DJ")]
+        [InlineData("Central Europe Standard Time", "Europe/Budapest", "\uFFFFNotValidRegion")]
         [InlineData("Central Europe Standard Time", "Europe/Prague", "CZ")]
         [InlineData("Central Europe Standard Time", "Europe/Ljubljana", "SI")]
         [InlineData("Central Europe Standard Time", "Europe/Bratislava", "SK")]
